rename modules and functions
