+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [0.1.2-dev]
 
 - reorganize tests directory
+- refactor module for batch effect correction. It is now named `pycombat`
+  (instead of `batch`), and the function to correct batch effect on microarray
+  data is now `pycombat_norm` (instead of `pycombat`).
 
 ## [0.1.1]
